Move zip and permission functions from local Utils to CommonUtils class in the shared project
